Create database folder + refactor to move database related files in there
